bootman_actions: mount downloaded rootfs as read-only
since the filesystem is journaling, every time we mount, it will change some metadata and write back changing the hash.
the detection of if the image should be redownloaded is done using md5, there is a md5 in http://filedump.furios.io/rootfs and it gets compared to the md5 of the locally downloaded rootfs.
if we mount as rw then this check will always return false and redownload since the checksum will always change during installation. mounting read only will prevent this since all writes are denied

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>